# internal/app/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rebecca API",
    "version": "0.2.0",
    "description": "Go-native Rebecca panel API. This document replaces the former FastAPI /docs page and is intended for dashboard, bot, and third-party client developers.\n\nAuthentication: protected routes use `Authorization: Bearer <token>`. The token can be the JWT returned by `POST /api/admin/token`; admin API keys are also accepted by the Go auth layer where bearer validation is used. Public subscription and home/docs routes explicitly set `security: []`.\n\nResponse compatibility: most endpoints intentionally keep the JSON shape expected by the current dashboard. Some schemas are documented as generic JSON while the migration to strict OpenAPI component schemas continues.\n\nMutation behavior: user, service, node, config, and host mutations are designed to be transactional. When runtime sync is required, Rebecca writes `node_operations` in the same transaction so failed queue writes roll back the database change."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Rebecca panel origin"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List users",
        "description": "List users with search, filters, sorting, owner/service scope, optional links, and traffic visibility enforcement.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an active authenticated admin. Standard admins are limited to users they own and services assigned to them; traffic fields may be zeroed when visibility is disabled.\n\nRoute details: Use repeated `username` or `filter` query parameters when needed. `links=true` asks the server to include subscription/config link fields, which costs extra DB/template work.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "name": "username",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": true
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "owner",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": true
          },
          {
            "name": "service_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "links",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getUsers"
      }
    },
    "/api/user": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create user",
        "description": "Create a user with required service assignment, permission checks, admin/service limits, credential normalization, and node operation enqueue.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an active authenticated admin. Standard admins are limited to users they own and services assigned to them; traffic fields may be zeroed when visibility is disabled.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.\n\nRoute details: Required business fields: `username` and `service_id`. `service_id` must point to an existing non-default service. Manual `inbounds` and `proxies` payloads are rejected; `credential_key` may be omitted, a Rebecca key, or a UUID that is normalized into the stored credential key.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/JSON"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postUser"
      }
    },
    "/api/v2/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create user",
        "description": "Versioned alias for creating a user.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an active authenticated admin. Standard admins are limited to users they own and services assigned to them; traffic fields may be zeroed when visibility is disabled.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.\n\nRoute details: This is the v2 alias of `/api/user`; it has the same validation and service-required behavior.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "postV2Users"
      }
    },
    "/api/user/{username}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user",
        "description": "Return a single user detail including subscription/config links and sanitized traffic values.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an active authenticated admin. Standard admins are limited to users they own and services assigned to them; traffic fields may be zeroed when visibility is disabled.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Username"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "getUserUsername"
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update user",
        "description": "Update user fields, service assignment, status, next plans, credential key, and node operation queue state.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an active authenticated admin. Standard admins are limited to users they own and services assigned to them; traffic fields may be zeroed when visibility is disabled.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.\n\nRoute details: Only send fields you want to change. `next_plan` was removed; use `next_plans`. If a legacy `setservice-<id>` inbound is supplied, Rebecca converts it to service assignment; ordinary inbound selection is rejected.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Username"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "putUserUsername"
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete user",
        "description": "Soft-delete a user and enqueue remove_user for runtime nodes.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an active authenticated admin. Standard admins are limited to users they own and services assigned to them; traffic fields may be zeroed when visibility is disabled.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Username"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "deleteUserUsername"
      }
    },
    "/api/v2/users/{username}": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update user",
        "description": "Versioned alias for updating a user.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an active authenticated admin. Standard admins are limited to users they own and services assigned to them; traffic fields may be zeroed when visibility is disabled.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.\n\nRoute details: Versioned update alias. Same validation and side effects as `PUT /api/user/{username}`.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Username"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "putV2UsersUsername"
      }
    },
...
    "/api/users/usage": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Users usage analytics",
        "description": "Return usage analytics grouped by admin for visible users.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an active authenticated admin. Standard admins are limited to users they own and services assigned to them; traffic fields may be zeroed when visibility is disabled.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Start"
          },
          {
            "$ref": "#/components/parameters/End"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getUsersUsage"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or Rebecca admin API key"
      }
    },
    "parameters": {
...
      "Offset": {
        "name": "offset",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 0
        },
        "description": "Zero-based row offset used for paginated list endpoints."
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1
        },
        "description": "Maximum number of rows to return. Large values may be capped by server-side settings."
      },
...
    },
...
  }
}